@@ -1,5 +1,5 @@
 import React, { CSSProperties } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import {
   TableContainer,
   Table,
@@ -67,18 +67,17 @@ const TableGenericCRUD = <DataTableInterface,>({
   canDelete,
   canEdit,
   canView,
-  viewUrl,
   onOpenEdit,
   onOpenDelete,
+  viewUrl,
 }: TableGenericProps<DataTableInterface>) => {
   const classes = useStyles();
-  const history = useHistory();
+  const navigate = useNavigate();
   const configMap: { [name: string]: TableHeader<DataTableInterface> } =
     headers.reduce((a, b) => {
       return { ...a, [b.field]: b };
     }, {});
   const headerFields = headers.map((header) => header.field);
-
   const renderHeaders = () => {
     return headers.map((header) => {
       const fieldName = setIfNotString(header.field);
@@ -203,7 +202,10 @@ const TableGenericCRUD = <DataTableInterface,>({
                     color="inherit"
                     aria-label="view"
                     className={classes.tableCellIcon}
-                    onClick={() => history.push(`${viewUrl}${row[idField]}`)}
+                    onClick={() =>
+                      (viewUrl && navigate(`${viewUrl}${row[idField]}`)) ||
+                      navigate(`${row[idField]}`)
+                    }
                     edge="start"
                   >
                     <FontAwesomeIcon icon={["fal", "eye"]} size="xs" />
