@@ -1,3 +1,5 @@
-const {useBabelRc, override} = require('customize-cra')
+const { useBabelRc, override } = require("customize-cra");
 
-module.exports = override(useBabelRc());
+module.exports = {
+  webpack: override(useBabelRc())
+};
