@@ -2,5 +2,7 @@ import { AxiosResponse } from "axios";
 import { mainAxiosClientManager } from "@clients/axios";
 import { API_ROUTES } from "@utils/constants";
 
+const { client } = mainAxiosClientManager;
+
 export const fetchPermissionCurrentUser = (): Promise<AxiosResponse> =>
-  mainAxiosClientManager.client.get(`${API_ROUTES.PERMISSION}current_user/`);
+  client.get(`${API_ROUTES.PERMISSION}current_user/`);
