@@ -1,16 +1,42 @@
 import axios, { AxiosInstance } from "axios";
 
+const { REACT_APP_API_URL = "http://127.0.0.1:8000/interaluse2021" } =
+  process.env;
+
 class AxiosClientManager {
+  private static instance: AxiosClientManager;
+
   client: AxiosInstance;
 
   tokenName: string;
 
-  constructor(baseUrl: string, tokenName: string) {
-    this.client = axios.create({
-      baseURL: baseUrl,
-    });
+  private constructor({
+    baseURL,
+    tokenName,
+  }: {
+    baseURL: string;
+    tokenName: string;
+  }) {
+    this.client = axios.create({ baseURL });
     this.tokenName = tokenName;
     this.loadToken();
+  }
+
+  public static getInstance({
+    baseURL,
+    tokenName,
+  }: {
+    baseURL: string;
+    tokenName: string;
+  }): AxiosClientManager {
+    if (!AxiosClientManager.instance) {
+      AxiosClientManager.instance = new AxiosClientManager({
+        baseURL,
+        tokenName,
+      });
+    }
+
+    return AxiosClientManager.instance;
   }
 
   addToken(token: string): void {
@@ -21,9 +47,13 @@ class AxiosClientManager {
   }
 
   removeToken(): void {
-    localStorage.removeItem(this.tokenName);
-    localStorage.removeItem("permissions");
-    delete this.client.defaults.headers.common.Authorization;
+    try {
+      localStorage.removeItem(this.tokenName);
+      localStorage.removeItem("permissions");
+      delete this.client.defaults.headers.common.Authorization;
+    } catch (error) {
+      console.log("The ERROR", error);
+    }
   }
 
   loadToken(): void {
@@ -36,9 +66,7 @@ class AxiosClientManager {
   }
 }
 
-export const mainAxiosClientManager = new AxiosClientManager(
-  process.env.NODE_ENV === "development"
-    ? "http://127.0.0.1:8000/interaluse2021"
-    : process.env?.REACT_APP_API_URL || "",
-  "main_token"
-);
+export const mainAxiosClientManager = AxiosClientManager.getInstance({
+  baseURL: REACT_APP_API_URL,
+  tokenName: "main_token",
+});
