@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import Button from "@material-ui/core/Button";
 import { DialogActions, Box, Grid } from "@material-ui/core";
 import { useForm } from "react-hook-form";
-import { BloodType, ExtendedUser, GenderType } from "@dbTypes/index";
+import { BloodType, ExtendedUser, GenderType } from "../../dto/index";
 import {
   DialogCreateEditBase,
   DialogCreateEditCRUDFormExtendedUser,
