@@ -1,5 +1,5 @@
 import React from "react";
-import { Escort } from "@dbTypes/escorts";
+import { Escort } from "../../dto/escorts";
 import { Grid, Typography, Box } from "@material-ui/core";
 import { INSTANCES_NAMES } from "@utils/constants";
 
